refactor: move Resolver to queries/ and flatten namespace
- Moved from services/ to queries/ directory
- Flattened DeliveryCriteria::Resolver to DeliveryCriteriaResolver
- Updated all references in controllers and interactors
- Simplified to V1 SIRET-only implementation
